chore: batch-landing fixes (per-arm GC store markers, index_set packed-loop split)
